+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [1.0.0] (Unreleased)
 
+- feat: cap and tail back-off execution crashes
 - feat: create configuration for job worker max crash count
 - feat: add `Ductwork::Workflow` alias of `Ductwork::Pipeline`
 - fix: skip "internal" errors when calculating advancement retry budget